+++ src/main/java/gr/iccs/smart/mobility/vehicle/VehicleInfoDTO.java
@@ -1,4 +1,4 @@
 package gr.iccs.smart.mobility.vehicle;
 
-public record VehicleInfoDTO(Float battery, Double latitude, Double longitude, VehicleStatus status) {
+public record VehicleInfoDTO(Double battery, Double latitude, Double longitude, VehicleStatus status) {
 }
